упорядочил структуру проекта
