+++ src/components/modal/avatar-change-modal/avatar-change-modal.hbs
