+++ src/core/renderer/renderer.spec.ts
@@ -1,5 +1,5 @@
 import { expect } from "expect";
-import { Block } from "../../core/block/block";
+import { Block } from "../block/block";
 import { renderer } from "./renderer";
 require("jsdom-global")('<div id="app"></div>', { url: "http://localhost" });
