+++ src/core/store/store.ts
@@ -1,4 +1,4 @@
-import { EventBus } from "../../services/event-bus/event-bus";
+import { EventBus } from "../event-bus/event-bus";
 
 export enum StoreEvents {
   Updated = "updated",
