--- src/layout/main.hbs
@@ -1,1 +1,0 @@
-<main class="main">{{{content}}}</main>
