--- src/layout/main.scss
@@ -1,5 +1,0 @@
-.main {
-  width: 100%;
-  height: 100%;
-  display: block;
-}
