+++ src/modules/profile/profile-edit/profile-edit.ts
@@ -1,13 +1,13 @@
 import tpl from "./profile-edit.hbs";
 import "./profile-edit.scss";
-import avatarUrl from "../../../static/Union.svg";
-import { Block } from "../../core/block/block";
-import { Button } from "../../components/button/button";
-import { Input } from "../../components/input/input";
-import { loginRegExp, nameRegExp, phoneRegExp } from "../../constants/regexps";
-import { getFormData, validationOnBlur } from "../../helpers/helpers";
-import { UserAPI } from "../../services/api/user";
-import { store, StoreEvents } from "../../core/store/store";
+import avatarUrl from "../../../../static/Union.svg";
+import { Block } from "../../../core/block/block";
+import { Button } from "../../../components/button/button";
+import { Input } from "../../../components/input/input";
+import { loginRegExp, nameRegExp, phoneRegExp } from "../../../constants/regexps";
+import { getFormData, validationOnBlur } from "../../../helpers/helpers";
+import { UserAPI } from "../../../services/api/user";
+import { store, StoreEvents } from "../../../core/store/store";
 
 const inputs = [
   new Input({
